backend/requirements.txt
# Firebase関連
firebase-admin>=6.2.0
google-cloud-firestore>=2.11.0

# FastAPI関連
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Vision AI関連
google-cloud-vision>=3.4.4
pillow>=10.0.0

# Vertex AI (Gemini)関連
google-cloud-aiplatform>=1.36.0
vertexai>=0.0.1

# ユーティリティ
python-jose[cryptography]>=3.3.0  # JWT認証用
passlib[bcrypt]>=1.7.4  # パスワードハッシュ化
aiofiles>=23.2.1  # 非同期ファイル操作 